Renamed classes and package
