+++ src/main/scala/beam/sim/BeamHelper.scala
@@ -9,7 +9,6 @@ import scala.collection.JavaConverters._
 import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
 import scala.concurrent.Await
-
 import beam.agentsim.agents.ridehail.{RideHailIterationHistory, RideHailSurgePricingManager}
 import beam.agentsim.events.handling.BeamEventsHandling
 import beam.analysis.ActivityLocationPlotter
@@ -26,9 +25,10 @@ import beam.sim.metrics.Metrics._
 import beam.sim.modules.{BeamAgentModule, UtilsModule}
 import beam.sim.population.PopulationAdjustment
 import beam.sim.ArgumentsParser.{Arguments, Worker}
+import beam.utils.csv.readers
+import beam.utils.scenario.matsim.BeamScenarioSource
 import beam.utils.{NetworkHelper, _}
 import beam.utils.scenario.{InputType, ScenarioLoader, ScenarioSource}
-import beam.utils.scenario.matsim.MatsimScenarioSource
 import beam.utils.scenario.urbansim.{CsvScenarioReader, ParquetScenarioReader, UrbanSimScenarioSource}
 import com.conveyal.r5.transit.TransportNetwork
 import com.fasterxml.jackson.databind.ObjectMapper
@@ -642,9 +642,9 @@ trait BeamHelper extends LazyLogging {
     if (src == "urbansim") {
       buildUrbansimScenarioSource(injector, beamConfig)
     } else if (src == "BeamCsv") {
-      new MatsimScenarioSource(
+      new BeamScenarioSource(
         scenarioFolder = beamConfig.beam.exchange.scenario.folder,
-        rdr = beam.utils.scenario.matsim.CsvScenarioReader
+        rdr = readers.BeamCsvScenarioReader
       )
     } else throw new NotImplementedError(s"ScenarioSource '$src' is not yet implemented")
   }
